Full name + Username Completed + Working on Level Bar
The full name and username are now completed and are shown in the profile.html page. First and last name are new data fields and are included in the register.html testing page. Currently, bar completeion is broken on the profile page along with some css from the bar that causes the profile picture to go away. Will fix next time.